* stable: (240 commits)
  (#22847) Restrict WMI group / user queries
  (maint) Explicit vcloud paths for pe nodes
  (#22324) Disable puppet listen on master/agent nodes
  (docs) Clarify Windows' handling of primary groups
  (maint) Coppy in the node configurations and pre-suite for pe test runs
  (maint) Remove stale vagrant support
  (#22324) Confine external_ca_support test not to run on PE
  (#22324) Loosen the host regex for package apply test on debian
  (maint) Fix default secs for destroy preserved hosts task
  (maint) Default xml output for beaker runs.
  (maint) Remove typo '3' in module_utils
  (maint) Check jenkins queueItem as well for a pending job
  (#22324) Puppet cert generate tests do not clear master ssl dir
  (#22324) Reports submission test checks puppetdb on pe runs
  (#22810) Fix rpm provider query method
  (#22804) Check for encoding support more robustly
  (#22324) Determine user/group from puppet configuration
  (#22324) Confine puppet manages own configuration test not to run on pe
  (#22324) Remove redundant without_installed_modules test
  (#22324) Fix modules/install/already_installed tests for pe
  ...

Conflicts:
	acceptance/tests/databinding/hiera/auto_lookup_for_class_parameters.rb
	acceptance/tests/databinding/hiera/bind_data_with_puppet_backend.rb

Resolved in favor of master having removed these tests in 75ab42.

+++ acceptance/tests/reports/submission.rb
@@ -1,13 +1,54 @@
 test_name "Report submission"
 
-testdir = master.tmpdir('report_submission')
+if master.is_pe?
+  require "time"
 
-with_master_running_on(master, "--reportdir #{testdir} --reports store --daemonize --dns_alt_names=\"puppet,$(facter hostname),$(facter fqdn)\" --autosign true") do
-  agents.each do |agent|
-    run_agent_on(agent, "-t --server #{master}")
+  def query_last_report_time_on(agent)
+    time_query_script = <<-EOS
+      require "net/http"
+      require "json"
 
-    on master, "grep -q #{agent} #{testdir}/*/*"
+      puppetdb_url = URI("http://localhost:8080/v3/reports")
+      puppetdb_url.query = URI.escape(%Q{query=["=","certname","#{agent}"]})
+      result = Net::HTTP.get(puppetdb_url)
+      json = JSON.load(result)
+      puts json.first["receive-time"]
+    EOS
+    puppetdb = hosts.detect { |h| h['roles'].include?('database') }
+    on(puppetdb, "#{master[:puppetbindir]}/ruby -e '#{time_query_script}'").output.chomp
   end
-end
 
-on master, "rm -rf #{testdir}"
+  last_times = {}
+
+  agents.each do |agent|
+    last_times[agent] = query_last_report_time_on(agent)
+  end
+
+  with_puppet_running_on(master, {}) do
+    agents.each do |agent|
+      on(agent, puppet('agent', "-t --server #{master}"))
+
+      current_time = Time.parse(query_last_report_time_on(agent))
+      last_time = Time.parse(last_times[agent])
+
+      assert(current_time > last_time, "Most recent report time #{current_time} is not newer than last report time #{last_time}")
+    end
+  end
+
+else
+
+  testdir = master.tmpdir('report_submission')
+
+  teardown do
+    on master, "rm -rf #{testdir}"
+  end
+
+  with_puppet_running_on(master, :main => { :reportdir => testdir, :reports => 'store' }) do
+    agents.each do |agent|
+      on(agent, puppet('agent', "-t --server #{master}"))
+
+      on master, "grep -q #{agent} #{testdir}/*/*"
+    end
+  end
+
+end
